// src/app/types.ts
export type TotalContributionPerYearTypes = {
  employerContribution: number;
  personalContribution: number;
};

export type PensionPotTypes = {
  jobStartingAge: number;
  retirementAge: number;
  employerContribution: number;
  personalContribution: number;
};

export type PensionFormValuesTypes = {
  employerContribution: number;
  personalContribution: number;
  retirementAge: number;
  desiredIncomePerYear: number;
};